app/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
requests>=2.31.0
PyPDF2>=3.0.1
python-dotenv>=1.0.1
mysql-connector-python
watchdog